+++ FYP/src/main/java/com/FYP/FYP/repository/NotificationRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
+public interface NotificationRepository extends JpaRepository<Notification, Integer> {
     List<Notification> findByUserAndIsReadFalse(User user);
 }
