+++ FYP/src/main/java/com/FYP/FYP/repository/TaskRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Long> {
-    List<Task> findByProjectId(Long projectId);
+public interface TaskRepository extends JpaRepository<Task, Integer> {
+    List<Task> findByProjectId(Integer projectId);
 }
